fix:move some duplicate code around
