+++ src/main/java/org/jahia/community/translation/assisted/service/impl/openai/OpenAITranslatorService.java
@@ -3,30 +3,25 @@ package org.jahia.community.translation.assisted.service.impl.openai;
 import com.openai.client.OpenAIClient;
 import com.openai.client.okhttp.OpenAIOkHttpClient;
 import com.openai.models.ChatModel;
-import com.openai.models.chat.completions.*;
+import com.openai.models.chat.completions.ChatCompletion;
+import com.openai.models.chat.completions.ChatCompletionCreateParams;
 import org.apache.commons.lang.StringUtils;
-import org.jahia.api.Constants;
 import org.jahia.community.translation.assisted.graphql.TranslatedField;
 import org.jahia.community.translation.assisted.service.AssistedTranslationResponse;
 import org.jahia.community.translation.assisted.service.TranslationServicesManager;
+import org.jahia.community.translation.assisted.service.TranslationServicesManagerImpl;
 import org.jahia.community.translation.assisted.service.TranslatorService;
 import org.jahia.community.translation.assisted.service.impl.TranslationData;
+import org.jahia.osgi.BundleUtils;
 import org.jahia.services.content.JCRNodeWrapper;
 import org.jahia.services.content.JCRSessionFactory;
 import org.jahia.services.content.JCRSessionWrapper;
-import org.jahia.services.content.nodetypes.ExtendedPropertyDefinition;
 import org.jahia.utils.LanguageCodeConverters;
 import org.json.JSONObject;
-import org.osgi.framework.BundleException;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.ConfigurationPolicy;
+import org.osgi.service.component.annotations.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.jcr.Property;
-import javax.jcr.PropertyIterator;
-import javax.jcr.PropertyType;
 import javax.jcr.RepositoryException;
 import java.text.MessageFormat;
 import java.util.List;
@@ -40,15 +35,17 @@ import static org.jahia.community.translation.assisted.AssistedTranslationsConst
 @Component(service = TranslatorService.class,
         property = {"service.translation.provider=openai","service.ranking=10.0"},
         configurationPid = SERVICE_CONFIG_FILE_NAME_OPENAI,
+        immediate = true,
         configurationPolicy = ConfigurationPolicy.REQUIRE)
 public class OpenAITranslatorService implements TranslatorService {
     private static final Logger logger = LoggerFactory.getLogger(OpenAITranslatorService.class);
-    private static final String SLASH = "/";
     private OpenAIClient openAIClient;
     private MessageFormat messageFormat;
     private ChatModel chatModel;
     private boolean available;
-    private Map<String, String> targetLanguages;
+
+    @Reference
+    private TranslationServicesManager translationServicesManager;
 
     @Override
     public String getProviderKey() {
@@ -66,7 +63,7 @@ public class OpenAITranslatorService implements TranslatorService {
             logger.warn("Missing configurations: {}", SERVICE_CONFIG_FILE_FULLNAME);
             return;
         }
-
+        logger.info("Activating OpenAI Translator Service");
         final String authKey = properties.getOrDefault(OPENAI_API_KEY, null);
         if (authKey == null) {
             available=false;
@@ -80,7 +77,6 @@ public class OpenAITranslatorService implements TranslatorService {
         if (properties.containsKey(TRANSLATION_OPENAI_MODEL)) {
             chatModel = ChatModel.of(properties.get(TRANSLATION_OPENAI_MODEL));
         }
-        targetLanguages = TranslationServicesManager.transformTargetLanguagesPropertiesToMap(properties);
         available=true;
     }
 
@@ -101,7 +97,7 @@ public class OpenAITranslatorService implements TranslatorService {
         final JCRNodeWrapper localizedNode = session.getNodeByIdentifier(node.getIdentifier());
         final TranslationData data = new TranslationData();
 
-        buildDataToTranslate(localizedNode, data, true);
+        translationServicesManager.buildDataToTranslate(localizedNode, data, true);
         // Build JSON to send to OpenAI, using data.getTexts() which is a map of key->text to translate, where key is the JCR path to the property (e.g. /a/title) and value is the text to translate (e.g. "Hello <b>world</b>")
 //        {
 //            "role": "user",
@@ -109,8 +105,8 @@ public class OpenAITranslatorService implements TranslatorService {
 //        }
         JSONObject json = new JSONObject(data.getTexts());
         // Use Chat Completions for traductions
-        String sourceLanguageMapped = targetLanguages.getOrDefault(sourceLanguage, sourceLanguage);
-        String targetLanguageMapped = targetLanguages.getOrDefault(targetLanguage, targetLanguage);
+        String sourceLanguageMapped = translationServicesManager.getTargetLanguages().getOrDefault(sourceLanguage, sourceLanguage);
+        String targetLanguageMapped = translationServicesManager.getTargetLanguages().getOrDefault(targetLanguage, targetLanguage);
         String systemPrompt = messageFormat.format(new Object[]{sourceLanguageMapped, targetLanguageMapped});
         ChatCompletionCreateParams params = ChatCompletionCreateParams
                 .builder()
@@ -135,62 +131,5 @@ public class OpenAITranslatorService implements TranslatorService {
         } else  {
             return List.of();
         }
-    }
-
-    private void buildDataToTranslate(JCRNodeWrapper node, TranslationData data, boolean forceTranslation) throws RepositoryException {
-        if (!isTranslatableNode(node)) {
-            return;
-        }
-
-        final PropertyIterator properties;
-        try {
-            properties = node.getProperties();
-        } catch (RepositoryException e) {
-            if (logger.isErrorEnabled()) {
-                logger.error("", e);
-            }
-            return;
-        }
-        while (properties.hasNext()) {
-            final Property property = properties.nextProperty();
-            if (isTranslatableProperty(property)) {
-                final String key = node.getPath() + SLASH + property.getName();
-                final String stringValue = StringUtils.trimToNull(property.getValue().getString());
-                if (forceTranslation || stringValue != null) {
-                    data.trackText(key, stringValue);
-                }
-            }
-
-        }
-    }
-
-    private boolean isTranslatableNode(JCRNodeWrapper node) {
-        try {
-            if (node.isNodeType(Constants.JAHIANT_PAGE)) {
-                return true;
-            }
-            if (node.isNodeType(Constants.JAHIANT_CONTENT)) {
-                return true;
-            }
-        } catch (RepositoryException e) {
-            if (logger.isErrorEnabled()) {
-                logger.error("", e);
-            }
-        }
-        return false;
-    }
-
-    private boolean isTranslatableProperty(Property property) {
-        final ExtendedPropertyDefinition definition;
-        try {
-            definition = (ExtendedPropertyDefinition) property.getDefinition();
-        } catch (RepositoryException e) {
-            if (logger.isErrorEnabled()) {
-                logger.error("", e);
-            }
-            return false;
-        }
-
-        return definition.isInternationalized() && !definition.isMultiple() && definition.getRequiredType() == PropertyType.STRING && !definition.isHidden() && !definition.isProtected();
     }
 }
